SDFSOF-101 - Fixed an error 500 on attempt to use notify_transaction_error/notify_transaction_expired

+++ platform/src/main/java/org/stellar/anchor/platform/action/NotifyTransactionErrorHandler.java
@@ -64,6 +64,8 @@ public class NotifyTransactionErrorHandler extends ActionHandler<NotifyTransacti
   @Override
   protected void updateTransactionWithAction(
       JdbcSepTransaction txn, NotifyTransactionErrorRequest request) {
-    transactionPendingTrustRepo.deleteById(txn.getId());
+    if (transactionPendingTrustRepo.existsById(txn.getId())) {
+      transactionPendingTrustRepo.deleteById(txn.getId());
+    }
   }
 }
